add directions for apple maps, add directionsMode, refactor
