+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"map_launcher","path":"/Users/alexmiller/projects/mine/plugins/map_launcher/","dependencies":[]}],"android":[{"name":"map_launcher","path":"/Users/alexmiller/projects/mine/plugins/map_launcher/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"map_launcher","dependencies":[]}],"date_created":"2020-07-25 12:15:03.403993","version":"1.17.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"map_launcher","path":"/Users/alexmiller/projects/mine/plugins/map_launcher/","dependencies":[]}],"android":[{"name":"map_launcher","path":"/Users/alexmiller/projects/mine/plugins/map_launcher/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"map_launcher","dependencies":[]}],"date_created":"2020-07-25 18:46:34.280856","version":"1.17.5"}
